requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
gpsd-py3==0.3.0
idna==3.10
linkify-it-py==2.0.3
mac-vendor-lookup==0.1.12
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
multidict==6.3.2
platformdirs==4.3.7
propcache==0.3.1
Pygments==2.19.1
requests==2.32.3
rich==14.0.0
scapy==2.6.1
textual==3.0.1
typing_extensions==4.13.1
uc-micro-py==1.0.3
urllib3==2.3.0
yarl==1.19.0
PyYAML~=6.0.2